+++ CHANGES.rst
@@ -9,6 +9,9 @@ in progress
 - Don't automatically start tailing the log file after "saraswati setup"
 - Add systemd path-activation file in order to start the Saraswati service only
   after all USB sound devices have been registered by watching ``/dev/bus/usb``.
+- Optionally use Ogg container format for audio recordings, by using the option
+  ``--container-format=ogg`` or the ``SARASWATI_CONTAINER_FORMAT`` environment
+  variable.
 
 
 2021-06-23 0.5.0